Created part of the Device class
I did the skeleton and some documentation (might not have been needed) and the logic for the spin and toString methods for the Device class, it still needs logic for peek and poke. If no one wants to finish those two methods then I'll have time later today to finish them. 
If anyone does want to finish it, be sure to use rotateOffset so we play around with indexes and don't have to keep swapping the data around. Also, someone might want to check the numbers I used in the code for spin, they seem arbitrary, but are they truly random for this implementation?